+++ src/test/java/lotto/domain/RandomNumberGeneratorTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
